+++ test/OpenTelemetry.Tests/Metrics/HostingMeterProviderBuilder.cs
@@ -1,0 +1,37 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+#if BUILDING_HOSTING_TESTS
+
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Diagnostics.Metrics;
+
+namespace OpenTelemetry.Metrics.Tests;
+
+public sealed class HostingMeterProviderBuilder : MeterProviderBuilderBase
+{
+    public HostingMeterProviderBuilder(IServiceCollection services)
+        : base(services)
+    {
+    }
+
+    public override MeterProviderBuilder AddMeter(params string[] names)
+    {
+        return this.ConfigureServices(services =>
+        {
+            foreach (var name in names)
+            {
+                // Note: The entire purpose of this class is to use the
+                // IMetricsBuilder API to enable Metrics and NOT the
+                // traditional AddMeter API.
+                services.AddMetrics(builder => builder.EnableMetrics(name));
+            }
+        });
+    }
+
+    public MeterProviderBuilder AddSdkMeter(params string[] names)
+    {
+        return base.AddMeter(names);
+    }
+}
+#endif
