+++ test/OpenTelemetry.Tests/Metrics/MetricTestsBase.cs
@@ -219,33 +219,6 @@ public abstract class MetricTestsBase
     }
 
 #if BUILDING_HOSTING_TESTS
-    public sealed class HostingMeterProviderBuilder : MeterProviderBuilderBase
-    {
-        public HostingMeterProviderBuilder(IServiceCollection services)
-            : base(services)
-        {
-        }
-
-        public override MeterProviderBuilder AddMeter(params string[] names)
-        {
-            return this.ConfigureServices(services =>
-            {
-                foreach (var name in names)
-                {
-                    // Note: The entire purpose of this class is to use the
-                    // IMetricsBuilder API to enable Metrics and NOT the
-                    // traditional AddMeter API.
-                    services.AddMetrics(builder => builder.EnableMetrics(name));
-                }
-            });
-        }
-
-        public MeterProviderBuilder AddSdkMeter(params string[] names)
-        {
-            return base.AddMeter(names);
-        }
-    }
-
     private sealed class MetricsSubscriptionManagerCleanupHostedService : IHostedService, IDisposable
     {
         private readonly object metricsSubscriptionManager;
